selenium/requirements.txt
python-owasp-zap-v2.4==0.0.14
selenium==3.14.1